pyproject.toml
[project]
name = "cacajogos"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-telegram-bot>=21.0",
    "httpx>=0.27",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9",
    "apscheduler>=3.10",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "python-amazon-paapi>=1.0",
    "beautifulsoup4>=4.12",
    "playwright>=1.48",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
